Place pulse subclasses in separate modules
Make ampx parameter of Pulse, not subclasses.
